requirements.txt
Flask-Cors
Flask-Login
Flask-PyMongo
Flask==2.3.3
pymongo==4.7.3
python-dotenv==1.0.1